refactor(framework): Extract Logger class to separate module
- Move Logger class from functional.py to a new logger.py module
- Add a new `warn()` method to the Logger class for handling warnings
- Update type hints in trainer.py to use Optional for package_func
